fix(metrics): retrieve metric type independently of its value
Previously, we used a metric's value to retrieve its type.
We can't do this as the Aggregate used to make the retrival is not
initialized.
It resulted in SIGSEGV for nil values (maps and fields).

The fix is simple: retrieve the type independently of the value,
via a new method reflectutil.PathResolver.ResolvePathType.

This commit also deletes the type metrics.Type
which became now useless.